requirements.txt
python-dotenv==1.0.1
pydantic==2.9.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2

# LangChain + backends
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.41
langchain-openai==0.1.23
langchain-text-splitters==0.2.2

# Vector DB (local)
chromadb==0.5.5
faiss-cpu==1.8.0.post1

# Parsing / ingest
pandas==2.2.2
beautifulsoup4==4.12.3

# Optional live tools (weather/search)
openmeteo-requests==1.2.0
requests-cache==1.2.1
retry-requests==2.0.0
numpy==1.26.4
tavily-python==0.5.0

# Tokenizer
tiktoken==0.7.0

# Local embeddings (HuggingFace)
sentence-transformers==2.2.2
torch>=2.0.0
